hello_world_keyboard/requirements.txt
pyserial>=3.4
keyboard>=0.11.0